+++ jira-rest-client-standalone/src/test/java/de/bschandera/MainTest.java
@@ -29,17 +29,20 @@ public class MainTest {
 
     @Before
     public void cleanJobsFolderAndVersionsFile() throws IOException {
-        Files.createDirectory(JOBS_PATH);
         Files.deleteIfExists(ROOT_PATH.resolve("versions.json"));
         Files.deleteIfExists(ROOT_PATH.resolve("config.json"));
-        Files.list(JOBS_PATH)
-                .forEach(file -> {
-                    try {
-                        Files.delete(file);
-                    } catch (IOException e) {
-                        throw new UncheckedIOException(e);
-                    }
-                });
+        if (Files.exists(JOBS_PATH)) {
+            Files.list(JOBS_PATH)
+                    .forEach(file -> {
+                        try {
+                            Files.delete(file);
+                        } catch (IOException e) {
+                            throw new UncheckedIOException(e);
+                        }
+                    });
+            Files.delete(JOBS_PATH);
+        }
+        Files.createDirectory(JOBS_PATH);
         Main.setJiraClient(null);
         Main.setPassword(new Password());
     }
@@ -48,14 +51,17 @@ public class MainTest {
     public static void cleanUpAfterwards() throws IOException {
         Files.deleteIfExists(ROOT_PATH.resolve("versions.json"));
         Files.deleteIfExists(ROOT_PATH.resolve("config.json"));
-        Files.list(JOBS_PATH)
-                .forEach(file -> {
-                    try {
-                        Files.delete(file);
-                    } catch (IOException e) {
-                        throw new UncheckedIOException(e);
-                    }
-                });
+        if (Files.exists(JOBS_PATH)) {
+            Files.list(JOBS_PATH)
+                    .forEach(file -> {
+                        try {
+                            Files.delete(file);
+                        } catch (IOException e) {
+                            throw new UncheckedIOException(e);
+                        }
+                    });
+            Files.delete(JOBS_PATH);
+        }
     }
 
     @Test
